Add basic multithreading support The Piano class can now calculate the next block of audio samples by running the FD scheme in two separate threads, one for each octave of the piano. This is more of a proof of concept and probably is not the best approach to multithreading, because the two threads are created and destroyed each time the method get_next_block_multithreaded() is invoked. It remains to be seen how much overhead is introduced by this approach. There are, however, tangible performance improvements. These new methods also return single-precision floats and accept a gain factor. The JUCE plugin has been updated to reflect all these changes.

+++ OpenPianoCore/string_hammer.h
@@ -384,7 +384,7 @@ struct String
 
         return current_sample;
     }
-    static drwav_uint64 save_to_wav(char* filename, double* sound, uint64_t duration_samples, bool normalize_output, bool destroy)
+    static drwav_uint64 save_to_wav(char* filename, float* sound, uint64_t duration_samples, bool normalize_output, bool destroy)
     {
         if(normalize_output)
         {
@@ -397,7 +397,7 @@ struct String
         format.format = DR_WAVE_FORMAT_IEEE_FLOAT; // <-- Any of the DR_WAVE_FORMAT_* codes.
         format.channels = 1;
         format.sampleRate = 48000;
-        format.bitsPerSample = sizeof (double)*8;
+        format.bitsPerSample = sizeof (float)*8;
         drwav_init_file_write(&wav, filename, &format, NULL);
         drwav_uint64 framesWritten = drwav_write_pcm_frames(&wav, duration_samples, sound);
         drwav_uninit(&wav);
